transcription/requirements.txt
openai-whisper
soundfile
tqdm
numpy